Remove 'render_data.render_data_utils' package and move its contents to 'renderer.render_data'
