--- src/main/scala/engine/render/renderer/render_data/render_data_util/doc.md
@@ -1,3 +1,0 @@
-# Overview
-
-This directory contains helper methods for the factory methods in the RenderData companion object.
